separate single problem loaders
